Give the Bootstrapper a basic GUI, set Ulterius Server to ANYCPU (was targeting x64 only), fix tray icon opening client no matter what.
