+++ Bootstrapper/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Costura.Fody" version="1.3.3.0" targetFramework="net461" developmentDependency="true" />
-  <package id="DotNetZip" version="1.10.1" targetFramework="net461" />
-  <package id="Fody" version="1.29.4" targetFramework="net461" developmentDependency="true" />
+  <package id="Costura.Fody" version="2.0.0-beta0018" targetFramework="net452" developmentDependency="true" />
+  <package id="DotNetZip" version="1.10.1" targetFramework="net452" />
+  <package id="Fody" version="1.30.0-beta01" targetFramework="net452" developmentDependency="true" />
 </packages>
