+++ RemoteTaskServer/FodyWeavers.xml
@@ -2,7 +2,6 @@
 <Weavers>
   <Costura>
     <ExcludeAssemblies>
-      UlteriusPluginBase
     </ExcludeAssemblies>
   </Costura>
